Refactor for moving intents down.
