+++ .vscode/settings.json
@@ -1,0 +1,19 @@
+{
+    "cSpell.words": [
+        "alexa",
+        "comparator",
+        "dump",
+        "esversion",
+        "livingroom",
+        "loadalbum",
+        "lordpengwin",
+        "playerid",
+        "playerindex",
+        "reprompt",
+        "squeezenode",
+        "squeezeserver",
+        "text",
+        "unsync",
+        "utterences"
+    ]
+}
