+++ intents/help.js
@@ -1,4 +1,4 @@
-const Intent = require("../intent");
+const Intent = require("./intent");
 const Utils = require("../utils");
 
 class Help extends Intent {
