+++ intents/sync.js
@@ -1,5 +1,5 @@
-const Intent = require("../intent");
 const Utils = require("../utils");
+const Intent = require("./intent");
 
 class Sync extends Intent {
     /**
@@ -22,37 +22,37 @@ class Sync extends Intent {
             // object for the second.
             player1 = this.findPlayerObject(squeezeserver, players, ((typeof intent.slots.FirstPlayer.value !== "undefined") && (intent.slots.FirstPlayer.value !== null) ? intent.slots.FirstPlayer.value : session.attributes.player));
             if (player1 === null) {
-                // Couldn"t find the player, return an error response
+                // Couldn't find the player, return an error response
                 console.log("Player not found: " + intent.slots.FirstPlayer.value);
                 callback(session.attributes, Utils.buildSpeechResponse(intentName, "Player not found", null, session.new));
             }
             session.attributes = { player: player1.name.toLowerCase() };
             player2 = null;
-            for (let pl of players) {
-                if (pl.name.toLowerCase() === this.normalizePlayer(intent.slots.SecondPlayer.value)) {
-                    player2 = p];
-            }
-        }
-        // If we found the target players, sync them
-        if (player1 && player2) {
-            console.log("Found players: %j and player2", player1, player2);
-            player1.sync(player2.playerindex, function(reply) {
-                if (reply.ok) {
-                    callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Synced " + player1.name + " to " + player2.name, null, session.new));
-                } else {
-                    console.log("Failed to sync %j", reply);
-                    callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Failed to sync players " + player1.name + " and " + player2.name, null, true));
+            for (let pl in players) {
+                if (players[pl].name.toLowerCase() === this.normalizePlayer(intent.slots.SecondPlayer.value)) {
+                    player2 = players[pl];
                 }
-            });
-        } else {
-            console.log("Player not found: ");
-            callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Player not found", null, session.new));
+            }
+            // If we found the target players, sync them
+            if (player1 && player2) {
+                console.log("Found players: %j and player2", player1, player2);
+                player1.sync(player2.playerindex, function(reply) {
+                    if (reply.ok) {
+                        callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Synced " + player1.name + " to " + player2.name, null, session.new));
+                    } else {
+                        console.log("Failed to sync %j", reply);
+                        callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Failed to sync players " + player1.name + " and " + player2.name, null, true));
+                    }
+                });
+            } else {
+                console.log("Player not found: ");
+                callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Player not found", null, session.new));
+            }
+        } catch (ex) {
+            console.log("Caught exception in syncPlayers %j for " + player1 + " and " + player2, ex);
+            callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Caught Exception", null, true));
         }
-    } catch (ex) {
-        console.log("Caught exception in syncPlayers %j for " + player1 + " and " + player2, ex);
-        callback(session.attributes, Utils.buildSpeechResponse("Sync Players", "Caught Exception", null, true));
     }
-}
 }
 
 module.exports = Sync.sync;
